+++ src/main/java/org/nullpointer/ratelimiter/model/state/TokenBucketState.java
@@ -1,19 +1,19 @@
 package org.nullpointer.ratelimiter.model.state;
 
 public class TokenBucketState implements RateLimitState {
-    private long availableTokens;
+    private double availableTokens;
     private long lastRefillTimestamp;
 
-    public TokenBucketState(long availableTokens, long lastRefillTimestamp) {
+    public TokenBucketState(double availableTokens, long lastRefillTimestamp) {
         this.availableTokens = availableTokens;
         this.lastRefillTimestamp = lastRefillTimestamp;
     }
 
-    public long getAvailableTokens() {
+    public double getAvailableTokens() {
         return availableTokens;
     }
 
-    public void setAvailableTokens(long availableTokens) {
+    public void setAvailableTokens(double availableTokens) {
         this.availableTokens = availableTokens;
     }
